moved settings back to settings.py rather than being a separate directory module
